+++ src/core/log/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.??.? (2021-??-??)
+
+#### :rocket: New Feature
+
+* Added a new middleware `ErrorsDeduplicatorMiddleware`
+
 ## v3.37.0 (2021-03-17)
 
 #### :boom: Breaking Change